Move email service to email module folder
