+++ modules/email/email.service.js
@@ -1,4 +1,4 @@
-const sgMail = require('../config/sendgrid');
+const sgMail = require('../../config/sendgrid');
 
 async function sendEmail({ to, subject, text, html }) {
   return sgMail.send({
